requirements.txt
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.1
idna==3.10
instagrapi==2.1.3
pillow==11.2.1
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
schedule==1.2.2
typing_extensions==4.13.2
urllib3==2.4.0
